fix: 온보딩 단계 DB 자료형 매핑 수정

+++ src/main/java/com/heapy/user/domain/User.java
@@ -16,7 +16,7 @@ public class User {
     private UUID userId;
 
     @Column(name = "onboarding_step", nullable = false)
-    private int onboardingStep;
+    private short onboardingStep;
 
     @Column(name = "onboarding_completed_at")
     private Instant onboardingCompletedAt;
